manage tags looks to be working

+++ mycelium/apps/mycelium_core/static/js/mycelium/manage_tags.js
@@ -15,6 +15,7 @@ manageTags.state = {};
 manageTags.objects = {};
 manageTags.actions = {};
 manageTags.handlers = {};
+manageTags.encyclopedia = {};
 manageTags.ui = {};
 
 manageTags.init = function() {
@@ -22,13 +23,19 @@ manageTags.init = function() {
 	manageTags.handlers.init();
 }
 
-manageTags.objects.tagSet = function(name, order ) {
+manageTags.encyclopedia.urls = {};
+
+manageTags.objects.tagSet = function(name, order, db_pk ) {
 	o = {};
-	o.pk = manageTags.state.tag_sets.length;
+	o.pk = manageTags.state.tagset_counter;
+	manageTags.state.tagset_counter ++;
 	o.name = name;
 	o.order = order;
 	o.is_deleted = false;
 	o.tags = [];
+	db_pk = (db_pk === undefined) ? false : db_pk;
+	o.db_pk = db_pk;
+
 
 	o.ui_element = function() {
 		return $("tagset[pk=" + this.pk + "]");
@@ -47,18 +54,40 @@ manageTags.objects.tagSet = function(name, order ) {
 		a.sort(manageTags.state.tag_and_tagset_sorter)
 		return a
 	}
+	o.set_order = function(order) {
+		this.order = order;
+	}
+	o.delete_from_page_and_state = function() {
+		for (var j in manageTags.state.tag_sets) {
+			if (manageTags.state.tag_sets[j].pk == this.pk) {
+				this.ui_element().remove();
+				manageTags.state.tag_sets.splice(j,1);
+			}
+		}
+	}
+	o.remove_tag = function(tag) {
+		for (var j in this.tags) {
+			if (tag.pk == this.tags[j].pk) {
+				this.tags.splice(j,1);
+				break;
+			}
+		}
+	}
 
 	return o;
 };
 
-manageTags.objects.tag = function(tagset, name, order, num_members ) {
+manageTags.objects.tag = function(tagset, name, order, num_members, db_pk ) {
 	o = {};
-	o.pk = tagset.tags.length;
+	o.pk = manageTags.state.tag_counter;
+	manageTags.state.tag_counter ++;
 	o.tagset = tagset;
 	o.name = name;
 	o.order = order;
 	o.num_members = num_members;
 	o.is_deleted = false;
+	db_pk = (db_pk === undefined) ? false : db_pk;
+	o.db_pk = db_pk;
 
 	o.ui_element = function() {
 		return $("li.tag[pk=" + this.pk + "]", this.ui_container());
@@ -69,6 +98,20 @@ manageTags.objects.tag = function(tagset, name, order, num_members ) {
 	o.delete = function() {
 		this.is_deleted = true;
 	}
+	o.set_order = function(order) {
+		this.order = order;
+	}
+	o.delete_from_page_and_state = function() {
+		for (var j in this.tagset.tags) {
+			if (manageTags.state.tag_sets[j].pk == this.pk) {
+				this.ui_element().remove();
+				this.tagset.tags.splice(j,1);
+			}
+		}
+	}
+	o.set_tagset = function(tagset) {
+		o.tagset = tagset;
+	}
 
 	return o;
 };
@@ -76,15 +119,48 @@ manageTags.objects.tag = function(tagset, name, order, num_members ) {
 
 // State
 manageTags.state.tag_sets = [];
+manageTags.state.tag_counter = 0;
+manageTags.state.tagset_counter = 0;
 manageTags.state.options = {};
-manageTags.state.options.save_timeout_length = 3000; //ms
+manageTags.state.options.save_timeout_length = 5000; //ms
+manageTags.state.save_in_progress = false;
+manageTags.state.save_queued = false;
 manageTags.state.get_tagset_from_click = function(target) {
-	return manageTags.state.tag_sets[parseInt(target.parents("tagset").attr("pk"),10)];
+	return manageTags.state.get_tagset_from_page_pk(parseInt(target.parents("tagset").attr("pk"),10));
 }
 manageTags.state.get_tag_from_click = function(target) {
-	var tagset = manageTags.state.get_tagset_from_click(target);
-	return tagset.tags[parseInt(target.parents(".tag").attr("pk"),10)];
+	var pk = parseInt(target.parents(".tag").attr("pk"),10);
+	return manageTags.state.get_tag_from_page_pk(pk);
 }
+
+manageTags.state.get_tagset_from_element = function(target){
+	return manageTags.state.get_tagset_from_page_pk(parseInt(target.attr("pk"),10));
+}
+manageTags.state.get_tag_from_element = function(target){
+	var pk = parseInt(target.attr("pk"),10);
+	return manageTags.state.get_tag_from_page_pk(pk);
+}
+
+manageTags.state.get_tagset_from_page_pk = function(pk) {
+	for (var j in manageTags.state.tag_sets) {
+	var tagset = manageTags.state.tag_sets[j];
+		if (tagset.pk == pk) {
+			return tagset;
+		}
+	}
+}
+manageTags.state.get_tag_from_page_pk = function(pk) {
+	for (var j in manageTags.state.tag_sets) {
+		var tagset = manageTags.state.tag_sets[j];
+		for (var i in tagset.tags) {
+			var tag = tagset.tags[i]
+			if (tag.pk == pk) {
+				return tag;
+			}
+		}
+	}
+}
+
 manageTags.state.tag_and_tagset_sorter = function(a, b) {
 	return a.order - b.order;
 }
@@ -95,6 +171,33 @@ manageTags.state.sorted_tagsets = function() {
 	}
 	a.sort(manageTags.state.tag_and_tagset_sorter)
 	return a
+}
+manageTags.state.full_data_dump = function() {
+	var data = {};
+	data.tag_sets = []
+	for (var i in manageTags.state.tag_sets) {
+		var tagset = manageTags.state.tag_sets[i];
+		var ts = {};
+		ts.name = tagset.name;
+		ts.order = tagset.order;
+		ts.is_deleted = tagset.is_deleted;
+		ts.tags = [];
+		ts.page_pk = tagset.pk;
+		ts.db_pk = tagset.db_pk;
+		for (var j in tagset.tags) {
+			var tag = tagset.tags[j];
+			var t = {};
+			t.name = tag.name;
+			t.order = tag.order;
+			t.is_deleted = tag.is_deleted;
+			t.db_pk = tag.db_pk;
+			t.page_pk = tag.pk;
+			ts.tags.push(t);
+		}
+
+		data.tag_sets.push(ts);
+	}
+	return {'data':JSON.stringify(data)}
 }
 
 
@@ -112,7 +215,7 @@ manageTags.handlers.tag_name_changed = function() {
 manageTags.handlers.delete_tag_clicked = function() {
 	var tag = manageTags.state.get_tag_from_click($(this));
 	tag.ui_element().addClass("pre_delete");
-	if (confirm("You sure?\n\nPress OK to delete this tag.\nPress Cancel to leave it in place.")) {
+	if (confirm("You sure?  \n\nThis will de-tag anyone with this tag, and delete the tag. It can't be undone.\n\nPress OK to delete this tag.\nPress Cancel to leave it in place.")) {
 		manageTags.actions.delete_tag(tag);
 	}
 	tag.ui_element().removeClass("pre_delete");
@@ -129,8 +232,8 @@ manageTags.handlers.add_tagset_clicked = function() {
 	return false;
 };
 manageTags.handlers.tagset_name_changed = function() {
-	var ts = manageTags.state.get_tagset_from_click($(this));
-	manageTags.actions.update_tag_obj(tag);
+	var tagset = manageTags.state.get_tagset_from_click($(this));
+	manageTags.actions.update_tagset_obj(tagset);
 	manageTags.actions.queue_save();
 
 };
@@ -149,24 +252,30 @@ manageTags.handlers.tagset_order_changed = function() {
 };
 
 manageTags.handlers.init = function(){
-	$( "tagsets" ).sortable({
+	manageTags.handlers.setup_sortable_handlers();
+	// .disableSelection();
+
+	$(".add_a_tag_btn").live("click",manageTags.handlers.add_tag_clicked);
+	$(".tag input").live("change",manageTags.handlers.tag_name_changed);
+	$(".delete_tag_btn").live("click",manageTags.handlers.delete_tag_clicked);
+
+	$(".add_a_category_btn").live("click",manageTags.handlers.add_tagset_clicked);
+	$("tagset .detail_header .generic_editable_field input").live("change",manageTags.handlers.tagset_name_changed);
+	$(".delete_tagset_btn").live("click",manageTags.handlers.delete_tagset_clicked);
+
+	$(".save_and_status_btn").live("click", manageTags.actions.queue_save);
+}
+manageTags.handlers.setup_sortable_handlers = function() {
+		$( "tagsets" ).sortable({
 					update: manageTags.handlers.tagset_order_changed,
+					// handle: '.tagset_drag_icon',
 				});
 	$( "ul.tags" ).sortable({
 		connectWith: "ul.tags",
 		update: manageTags.handlers.tag_order_changed,
+		// handle: '.tag_drag_icon',
 	});
-	// .disableSelection();
-
-	$(".add_a_tag_btn").live("click",manageTags.handlers.add_tag_clicked);
-	$(".tag input[name$=-name").live("change",manageTags.handlers.tag_name_changed);
-	$(".delete_tag_btn").live("click",manageTags.handlers.delete_tag_clicked);
-
-	$(".add_a_category_btn").live("click",manageTags.handlers.add_tagset_clicked);
-	$("tagset input[name$=-name").live("change",manageTags.handlers.tagset_name_changed);
-	$(".delete_tagset_btn").live("click",manageTags.handlers.delete_tagset_clicked);
 }
-
 
 // Actions
 manageTags.actions.add_tag = function(tagset) {
@@ -174,6 +283,7 @@ manageTags.actions.add_tag = function(tagset) {
 	manageTags.ui.render_tagset(tagset);
 	manageTags.actions.sort_tags_and_tagsets();
 	manageTags.actions.queue_save();
+	manageTags.handlers.setup_sortable_handlers();
 };
 manageTags.actions.update_tag_obj = function(tag) {
 	tag.name = $(".tag_name input[name$=-name]",tag.ui_element()).val();
@@ -189,10 +299,11 @@ manageTags.actions.add_tagset = function() {
 	manageTags.ui.render_tagset(tagset);
 	manageTags.actions.sort_tags_and_tagsets();
 	manageTags.actions.queue_save();
+	manageTags.handlers.setup_sortable_handlers();
 
 };
 manageTags.actions.update_tagset_obj = function(tagset) {
-	tagset.name = $(".detail_header input[name$=-name]",tag.ui_element()).val();
+	tagset.name = $(".detail_header input[name$=-name]",tagset.ui_element()).val();
 };
 manageTags.actions.delete_tagset = function(tagset) {
 	tagset.delete();
@@ -200,29 +311,91 @@ manageTags.actions.delete_tagset = function(tagset) {
 	manageTags.actions.queue_save();
 };
 manageTags.actions.queue_save = function() {
-	clearTimeout(manageTags.state.save_timeout)
-	manageTags.state.save_timeout = setTimeout(manageTags.actions.save_state, manageTags.state.options.save_timeout_length);
+	manageTags.state.save_queued = true;
+	if (!manageTags.state.save_in_progress) {
+		manageTags.ui.show_saving();
+		clearTimeout(manageTags.state.save_timeout)
+		manageTags.state.save_timeout = setTimeout(manageTags.actions.save_state, manageTags.state.options.save_timeout_length);
+	}
 };
 manageTags.actions.save_state = function() {
-	console.log("saving");
+	manageTags.state.save_queued = false;
+	manageTags.state.save_in_progress = true;
+	
+	$.ajax({
+		url: manageTags.encyclopedia.urls.save_tags_and_tagsets,
+		type: "POST",
+		dataType: "json",
+		data: manageTags.state.full_data_dump(),
+		success: function(json) {
+			// delete the things that were deleted.
+			for (var j in json.created_tagsets) {
+				var ts = json.created_tagsets[j];
+				var tagset = manageTags.state.get_tagset_from_page_pk(ts.page_pk)
+				tagset.db_pk = ts.db_pk;
+			}
+    		for (var j in json.created_tags) {
+	    		var t = json.created_tags[j];
+	    		var tag = manageTags.state.get_tag_from_page_pk(t.page_pk)
+	    		tag.db_pk = t.db_pk;
+	    	}
+    		for (var j in json.deleted_tagsets) {
+	    		var ts = json.deleted_tagsets[j];
+	    		var tagset = manageTags.state.get_tagset_from_page_pk(ts.page_pk)
+	    		// make sure it's still gone
+	    		if (tagset.is_deleted) {
+	    			tagset.delete_from_page_and_state()	
+	    		}
+	    	}
+    		for (var j in json.deleted_tags) {
+	    		var t = json.deleted_tags[j];
+	    		var tag = manageTags.state.get_tag_from_page_pk(t.page_pk)
+	    		if (tag.is_deleted) {
+	    			tag.delete_from_page_and_state()	
+	    		}
+	    	}
+
+			if (manageTags.state.save_queued) {
+				manageTags.actions.save_state();
+			} else {
+				manageTags.state.save_in_progress = false;
+				manageTags.ui.show_saved();
+			}
+			
+		}
+	});
 };
 manageTags.actions.sort_tags_and_tagsets = function (){
 	var ts_counter = 0;
+	var tag_counter = 0;
 	$("tagset").each(function(){
 		var ts_ele = $(this);
-		var tagset = manageTags.state.get_tagset_from_click(ts_ele.children(":first"));
-		tagset.order = ts_counter;
-
-		var tag_counter = 0;
-		$(".tag", ts_ele).each(function() {
-			var tag_ele = $(this);
-			var tag = manageTags.state.get_tag_from_click(tag_ele.children(":first"));
-			tag.order = tag_counter;
-			tag_counter ++;
-		})
+		manageTags.state.get_tagset_from_element(ts_ele).set_order(ts_counter);
 		ts_counter++;
 	});
+	$("li.tag").each(function() {
+		var tag_ele = $(this);
+		var tag = manageTags.state.get_tag_from_element(tag_ele);
+		var tagset = manageTags.state.get_tagset_from_click(tag_ele);
+		tag.set_order(tag_counter);
+		if (tag.tagset != tagset) {
+			tag.tagset.remove_tag(tag);
+			tag.set_tagset(tagset);	
+			tagset.tags.push(tag);
+		}
+		
+		tag_counter ++;
+	});
+
+	// Update any tagsets.
+	for (var j in manageTags.state.tag_sets) {
+		var tagset = manageTags.state.tag_sets[j];
+		for (var i in tagset.tags) {
+			
+		}
+	}
 }
+
 
 manageTags.ui.init = function() {
 	manageTags.ui.render_full_ui();
@@ -230,13 +403,13 @@ manageTags.ui.init = function() {
 manageTags.ui.render_tag = function(tag) {
 	var str = "";
 	if (!tag.is_deleted) {
-		str += '<div><table class="tag_list striped">';
+		str += '<table class="tag_list striped">';
 		str += '	<tr class="striped_row tag_row">';
 		str += '		<td class="tag_name"><div class="tag_drag_icon"></div><span class="generic_editable_field"><span class="edit_field"><input type="text" name="TAG-'+tag.pk+'-name" value="'+tag.name+'" maxlength="250" /></span></span></td>';
 		str += '		<td class="num_people"><span class="count">'+tag.num_members+'</span> '+ ((tag.num_members==1)?'person':'people') +'</td>';
 		str += '		<td class="tag_actions"><a href="#" class="delete_tag_btn mycelium_btn mycelium_delete_btn mycelium_active_grey mycelium_small_btn">Delete Tag</a></td>';
 		str += '	</tr>';
-		str += '</table></div>';
+		str += '</table>';
 
 		if (!tag.ui_element().length) {
 			tag.ui_container().append('<li class="tag" pk="'+tag.pk+'"></li>')
@@ -291,69 +464,12 @@ manageTags.ui.render_full_ui = function() {
 };
 
 
-// function bind_tags_events() {
-// 	$(".delete_tagset_btn").die("click").live("click",delete_category);
-// 	$(".delete_tag_btn").die("click").live("click",delete_tag);
-// 	$(".add_a_tag_btn").die("click").live("click",add_a_tag);
-// 	$(".add_a_category_btn").die("click").live("click",add_a_category);
-// }
+manageTags.ui.show_saving = function() {
+	$(".save_and_status_btn").html("Saving...");
+	$(".last_save_time").html("")
+}
 
-// function process_link_via_json(t) {
-// 	setTimeout(function(){
-// 	if ($("#basic_info_form").hasClass("dirty")) {
-// 		$("#basic_info_form").ajaxSubmit({
-// 			'async':false
-// 		});		
-// 	}
-// 	$.ajax({
-// 		url: $(t).attr("href"),
-// 		type: "POST",
-// 		dataType: "json",
-// 		success: function(json) {
-// 			process_fragments_and_rebind_tags_form(json);
-// 		}
-// 	});
-// 	},10);
-// }
-
-// function delete_category() {
-// 	var ts = $(this).parents("tagset");
-// 	var tagset_row = $(".detail_header",ts);
-// 	var tag_rows = $(".tag_row",ts);
-// 	var ts_add = $(".add_a_tag_btn",ts);
-// 	tagset_row.addClass("pre_delete");
-// 	tag_rows.addClass("pre_delete");
-// 	ts_add.addClass("pre_delete")
-// 	if (confirm("Hold on there.\n\nThis will delete the entire tag category, including all tags in it!\n\nThis action can not be undone.\n\nPress OK to delete this tag category.\nPress Cancel to leave it intact.")) {
-// 		process_link_via_json($(this));
-// 	} else {
-// 		tagset_row.removeClass("pre_delete");
-// 		tag_rows.removeClass("pre_delete");
-// 		ts_add.removeClass("pre_delete");
-// 	}
-// 	return false;
-// }
-// function delete_tag() {
-// 	var tag_row = $(this).parents(".tag_row");
-// 	tag_row.addClass("pre_delete");
-// 	if (confirm("You sure?\n\nPress OK to delete this tag.\nPress Cancel to leave it in place.")) {
-// 		process_link_via_json($(this));
-// 	} else {
-// 		tag_row.removeClass("pre_delete");
-// 	}
-// 	return false;
-// }
-// function add_a_tag() {
-// 	process_link_via_json($(this));
-// 	return false;
-// }
-// function add_a_category() {
-// 	process_link_via_json($(this));
-// 	return false;
-// }
-
-// function process_fragments_and_rebind_tags_form(json) {
-//     $.Mycelium.fragments.process_fragments_from_json(json);
-//     $.Mycelium.update_stripes();
-//     bind_tags_events();
-// }
+manageTags.ui.show_saved = function() {
+	$(".save_and_status_btn").html("Saved");
+	$(".last_save_time").html("Saved a few seconds ago.")
+}
